+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v0.1.0 - 2021-04-05
+~~~~~~~~~~~~~~~~~~~
+
+* `cjwpandasmodule.convert`: convert Pandas to Arrow (and back)
+* `cjwpandasmodule.validate`: allow `period[D]` values (for forward
+  compatibility -- Workbench does not support this as of 2021-04-05).
+* 100% unit-test coverage
+
 v0.0.2 - 2021-01-22
 ~~~~~~~~~~~~~~~~~~~